Fix tab order in DisplayRangeSelector and RangeSelector
The changes modify the order in which controls are added to their
parent controls to ensure a logical tab navigation sequence. In
DisplayRangeSelector, the `time` control is now added before the
`signal` control. Similarly, in RangeSelector, the `from` control
is now added before the `to` control. This ensures that the tab
index of the `time` and `from` controls is less than that of the
`signal` and `to` controls, respectively.